requirements.txt
flask_wtf
sqlalchemy
flask-login
psycopg2
email_validator
flask_socketio
flask_sqlalchemy